# Obsevatoire_de_recherche/src/Backend/laravel10/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for my application.",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/api/usecases/search/Projects": {
            "get": {
                "tags": [
                    "Recherche"
                ],
                "summary": "Rechercher des projets",
                "operationId": "7072072be454cc8f86474555765e23dc",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Texte de recherche pour les projets",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des projets correspondants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TblProjet"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide"
                    }
                }
            }
        },
        "/api/usecases/search/Documents": {
            "get": {
                "tags": [
                    "Recherche"
                ],
                "summary": "Rechercher des documents",
                "operationId": "ef0d4b6124328af4904513127d5bfa63",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Texte de recherche pour les documents",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des documents correspondants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TblDocument"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide"
                    }
                }
            }
        },
        "/api/usecases/search/categories": {
            "get": {
                "tags": [
                    "Recherche"
                ],
                "summary": "Rechercher des catégories",
                "operationId": "dabf2f02c4e7511d8b6de6d3a060643b",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Texte de recherche pour les catégories",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des catégories correspondantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TblCategorie"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "TblCategorie": {
                "title": "TblCategorie",
                "required": [
                    "nom_cat",
                    "descript_cat"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the category",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom_cat": {
                        "description": "Name of the category",
                        "type": "string"
                    },
                    "descript_cat": {
                        "description": "Description of the category",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TblDocument": {
                "title": "TblDocument",
                "required": [
                    "nom_doc",
                    "lien_doc",
                    "type_doc",
                    "resume",
                    "tbl_projet_id"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the document",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nom_doc": {
                        "description": "Name of the document",
                        "type": "string"
                    },
                    "lien_doc": {
                        "description": "Link to the document",
                        "type": "string"
                    },
                    "type_doc": {
                        "description": "Type of the document (PDF, WORD, POWEPOINT)",
                        "type": "string"
                    },
                    "resume": {
                        "description": "Summary of the document",
                        "type": "string"
                    },
                    "tbl_projet_id": {
                        "description": "ID of the related project",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "TblProjet": {
                "title": "TblProjet",
                "required": [
                    "titre_projet",
                    "descript_projet",
                    "tbl_niveau_id",
                    "user_id",
                    "tbl_categorie_id"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the projet",
                        "type": "integer",
                        "format": "int64"
                    },
                    "titre_projet": {
                        "description": "Title of the projet",
                        "type": "string"
                    },
                    "descript_projet": {
                        "description": "Description of the projet",
                        "type": "string"
                    },
                    "tbl_niveau_id": {
                        "description": "The ID of the associated niveau",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "The ID of the associated user",
                        "type": "integer",
                        "format": "int64"
                    },
                    "tbl_categorie_id": {
                        "description": "The ID of the associated categorie",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}